bot/python/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.12.2
cachetools==5.3.1
discord.py==2.3.0
discord-ext-menus==1.1
emoji==2.5.0
html2text==2020.1.16
lxml==4.9.2
numpy==1.24.3
python_dotenv==0.21.1
requests==2.25.1
scikit_image==0.21.0
scikit-learn==1.2.2